+++ lib/dockerfile_distro.mli
@@ -28,7 +28,7 @@ type t = [
   | `Debian of [ `Stable | `Testing | `Unstable ]
   | `Fedora of [ `V21 | `V22 | `V23 ]
   | `OracleLinux of [ `V7 ]
-  | `Ubuntu of [ `V14_04 | `V15_04 | `V15_10 ]
+  | `Ubuntu of [ `V12_04 | `V14_04 | `V15_04 | `V15_10 | `V16_04 ]
 ] with sexp 
 (** Supported Docker container distributions *)